+++ lru_trait/src/lib.rs
@@ -1,3 +1,3 @@
 pub trait LRU<T> {
-    fn rec_access(&mut self, val: T) -> Option<u32>;
+    fn rec_access(&mut self, val: T) -> Option<usize>;
 }
